app/requirements.txt
python-dotenv
requests
bs4
ollama
selenium
redis
colorist